backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic
pytest
httpx
pytest-cov
